DeepTrafficNet.js
This is an initial architecture for the DeepTraffic Network. This Architecture uses 2 hidden layers. One with 1000 neurons which are connected to the input and the output has 3000 neurons. The activation function used was tanh for both layers. The Gamma parameter which signifies how much the net is sensitive to future instances. So I have set this parameter really high to 0.7. There is further room for improvement with even higher gamma value. The experience size was 1500 and Threshold was only 500: Meaning we wait for 500 samples before we make decisions. The learning rate used was 0.001 and Number of iterations were only 5000. There is more room for improvement with higher gamma and number of iteration values. The Score obtained by this network was